requirements.txt
fastapi==0.78.0
uvicorn==0.17.6
FastAPI-SQLAlchemy==0.2.1
strawberry-graphql==0.114.3
psycopg2-binary==2.9.3
